% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.survData.R
\name{as.survData}
\alias{as.survData}
\title{Cast for classes derived from \code{survData}}
\usage{
as.survData(x)
}
\arguments{
\item{x}{an S3 object of a class derived from \code{survData}}
}
\description{
Use this function to access \code{survData} methods on an object of a 
derived class (e.g. \code{reproData})
}
\examples{
data(zinc)
x <- reproData(zinc)

# Compare
plot(x)

#and
plot(as.survData(x))

}
\seealso{
\code{\link{reproData}} and \code{\link{survData}}
}

