% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\docType{data}
\name{cadmium2}
\alias{cadmium2}
\title{Reproduction and survival data sets for \emph{Lymnaea stagnalis} exposed to cadmium during 28
days}
\format{
A data frame with 612 observations of the following five variables:
\describe{ \item{\code{replicate}}{A vector of class \code{numeric} with the
replicate code (\code{1} to \code{36}).} \item{\code{conc}}{A vector of
class \code{integer} with the cadmium concentrations in \eqn{\mu g.L^{-1}}.}
\item{\code{time}}{A vector of class \code{integer} with the time points
(in days from the beginning of the experiment \eqn{t = 0}).}
\item{\code{Nsurv}}{A vector of class \code{integer} with the number of
alive individuals at each time point for each concentration and each
replicate.} \item{\code{Nrepro}}{A vector of class \code{integer} with the
number of clutches at each time point for each concentration and each
replicate.} }
}
\usage{
data(cadmium2)
}
\description{
Reproduction and survival data sets of chronic laboratory toxicity tests with
snails (\emph{Lymnaea stagnalis}) exposed to six concentrations of cadmium
during 28 days. Six concentrations were tested, with six replicates per
concentration. Each replicate contained five organisms. Reproduction and
survival were monitored at 17 time points.
}
\references{
Ducrot, V., Askem, C., Azam, D., Brettschneider, D., Brown,
R., Charles, S., Coke, M., Collinet, M., Delignette-Muller, M.L.,
Forfait-Dubuc, C., Holbech, H., Hutchinson, T., Jach, A., Kinnberg, K.L.,
Lacoste, C., Le Page, G., Matthiessen, P., Oehlmann, J., Rice, L.,
Roberts, E., Ruppert, K., Davis, J.E., Veauvy, C., Weltje, L., Wortham, R.
and Lagadic, L. (2014)
Development and validation of an OECD reproductive toxicity test guideline with
the pond snail Lymnaea stagnalis (Mollusca, Gastropoda),
\emph{Regulatory Toxicology and Pharmacology}, 70(3), 605-14.

Charles, S., Ducrot, V., Azam, D., Benstead, R., Brettschneider, D., De Schamphelaere, K.,
Filipe Goncalves, S., Green, J.W., Holbech, H., Hutchinson, T.H., Faber, D., Laranjeiro, F.,
Matthiessen, P., Norrgren, L., Oehlmann, J., Reategui-Zirena, E., Seeland-Fremer, A., Teigeler, M.,
Thome, J.P., Tobor Kaplon, M., Weltje, L., Lagadic, L. (2016) 
Optimizing the design of a reproduction toxicity test with the pond snail Lymnaea stagnalis,
\emph{Regulatory Toxicology and Pharmacology}, vol. 81 pp.47-56.
}
\keyword{data}
\keyword{set}
