% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\docType{data}
\name{chlordan}
\alias{chlordan}
\title{Reproduction and survival data sets for \emph{Daphnia magna} exposed to
chlordan during 21 days}
\format{
A data frame with 1320 observations of the following five variables:
\describe{ \item{\code{replicate}}{A vector of class \code{numeric} with
the replicate code (\code{1} to \code{60}).} \item{\code{conc}}{A vector of
class \code{numeric} with the chlordan concentrations in \eqn{\mu
g.L^{-1}}.} \item{\code{time}}{A vector of class \code{integer} with the
time points (in days from the beginning of the experiment \eqn{t = 0}).}
\item{\code{Nsurv}}{A vector of class \code{integer} with the number of
alive individuals at each time point for each concentration and each
replicate.} \item{\code{Nrepro}}{A vector of class \code{integer} with the
number of offspring at each time point for each concentration and each
replicate.} }
}
\usage{
data(chlordan)
}
\description{
Reproduction and survival data sets of chronic laboratory toxicity tests with
\emph{Daphnia magna} freshwater invertebrate exposed to six concentrations
of one organochlorine insecticide (chlordan) during 21 days. Six concentrations were
tested, with 10 replicates per concentration. Each replicate contained one
organism. Reproduction and survival were monitored at 22 time points.
}
\references{
Manar, R., Bessi, H. and Vasseur, P. (2009) Reproductive effects
and bioaccumulation of chlordan in Daphnia magna, \emph{Environmental
Toxicology and Chemistry}, 28, 2150-2159.
}
\keyword{data}
\keyword{set}
