% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\docType{data}
\name{FOCUSprofile}
\alias{FOCUSprofile}
\title{A simulated exposure profile with 11641 time points.}
\format{
A data frame with 11641 observations on the following two variables:
\describe{ \item{\code{time}}{A vector of class \code{numeric}.}
 \item{\code{conc}}{A vector of class \code{numeric} with exposure concentrations.} and 
 \item{\code{replicate}}{A vector of class \code{factor}.} }
}
\usage{
data(FOCUSprofile)
}
\description{
Exposure profile of 11641 time points used for prediction.
}
\keyword{data}
\keyword{set}
