% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survData.R
\name{is_exposure_constant}
\alias{is_exposure_constant}
\title{Test in a well-formed argument to function 'survData' if the concentration
is constant and different from \code{NA} for each replicate (each time-serie)}
\usage{
is_exposure_constant(x)
}
\arguments{
\item{x}{an object of class \code{data.frame}}
}
\value{
a boolean \code{TRUE} if concentration in \code{replicate} is constant,
 or \code{FALSE} if the concentration in at least one of the replicates is time-variable,
 and/or if \code{NA} occures.
}
\description{
Test in a well-formed argument to function 'survData' if the concentration
is constant and different from \code{NA} for each replicate (each time-serie)
}
\examples{

# (1) Load the survival data set and test if concentration in replicates is constant
data("propiconazole")
is_exposure_constant(propiconazole)
is_exposure_constant(survData(propiconazole))

 # (1) Load the survival data set and test if concentration in replicates is constant
data("propiconazole_pulse_exposure") 
is_exposure_constant(propiconazole_pulse_exposure)

}
