% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCx.R
\name{LCx}
\alias{LCx}
\title{Predict \eqn{X}\% Lethal Concentration at the maximum time point (default).}
\usage{
LCx(object, ...)
}
\arguments{
\item{object}{An object used to select a method}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Predict median and 95\% credible interval of the x\% Lethal Concentration.
}
