% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{make_code_table}
\alias{make_code_table}
\title{Make a hash table for code.}
\usage{
make_code_table(code, from = 1, to = 2)
}
\arguments{
\item{code}{Mapping from characters to code words. Each element is a vector
of two strings.}

\item{from}{Index of the character in the element of \code{code}.}

\item{to}{Index of the code word in the element of \code{code}.}
}
\value{
The hash table for code.
}
\description{
Make a hash table for code.
}
\examples{
morse:::make_code_table(list(c("a", "1"), c("b", "2")))
morse:::make_code_table(list(c("a", "1"), c("b", "2")), from = 2, to = 1)
}
