% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{make_decode}
\alias{make_decode}
\title{Make a decode function.}
\usage{
make_decode(table, pre = identity)
}
\arguments{
\item{table}{Hash table for looking up characters for code words.}

\item{pre}{A function to pre-process the input.}
}
\value{
A function that takes an encoded string and return the original string.
}
\description{
Make a decode function.
}
\examples{
table <- morse:::make_code_table(list(c("a", "1"), c("b", "2")), from = 2, to = 1)
decode <- make_decode(table)
decode("1/2", sep = "/")
}
