% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{make_encode}
\alias{make_encode}
\title{Make an encode function.}
\usage{
make_encode(table, pre = identity)
}
\arguments{
\item{table}{Hash table for looking up code words.}

\item{pre}{A function to pre-process the input.}
}
\value{
A function that takes a string and return the encoded string.
}
\description{
Make an encode function.
}
\examples{
table <- morse:::make_code_table(list(c("a", "1"), c("b", "2")))
encode <- make_encode(table)
encode("ab")
}
