% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFx_ode.survFit.R
\name{MFx_ode}
\alias{MFx_ode}
\title{Predict x\% Multiplication Factor method for \code{survFit} objects}
\usage{
MFx_ode(object, ...)
}
\arguments{
\item{object}{an object used to select a method \code{ppc}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
This is a \code{method} to replace function \code{MFx} used on \code{survFit}
object when computing issues happen. \code{MFx_ode} uses the \code{deSolve}
library to improve robustness. However, time to compute may be longer.
}
