% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R, R/modelData.survDataCstExp.R,
%   R/modelData.survDataVarExp.R
\name{modelData}
\alias{modelData}
\alias{modelData.survDataCstExp}
\alias{modelData.survDataVarExp}
\title{Create a list giving data to use in Bayesian inference.}
\usage{
modelData(x, ...)

\method{modelData}{survDataCstExp}(x, model_type = NULL, ...)

\method{modelData}{survDataVarExp}(x, model_type = NULL, extend_time = 100, ...)
}
\arguments{
\item{x}{An object of class \code{survData}}

\item{\dots}{Further arguments to be passed to generic methods}

\item{model_type}{TKTD GUTS model type ('SD' or 'IT')}

\item{extend_time}{Number of for each replicate used for linear 
interpolation (comprise between time to compute and fitting accuracy)}
}
\value{
A list for parameterization of priors for Bayesian inference.

A list for parameterization of priors for Bayesian inference.

A list for parameterization of priors for Bayesian inference.
}
\description{
Create a data set to analyse a \code{survDataCstExp} object.

Create a data set to analyse a \code{survDataVarExp} object.
}
