% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.survDataVarExp.R
\name{modelData.survDataVarExp}
\alias{modelData.survDataVarExp}
\title{Create a data set to analyse a \code{survDataVarExp} object.}
\usage{
\method{modelData}{survDataVarExp}(x, model_type = NULL, extend_time = 100, ...)
}
\arguments{
\item{x}{An object of class \code{survData}}

\item{model_type}{TKTD GUTS model type ('SD' or 'IT')}

\item{extend_time}{Number of for each replicate used for linear 
interpolation (comprise between time to compute and fitting accuracy)}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
A list for parameterization of priors for Bayesian inference.
}
\description{
Create a data set to analyse a \code{survDataVarExp} object.
}
