% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{morse}
\alias{morse}
\title{Encode a string with Morse code.}
\usage{
morse(string, sep = "/")
}
\arguments{
\item{string}{A character vector of length 1 to encode.}

\item{sep}{Separator between code words.}
}
\value{
Morse encoded string.
}
\description{
Encode a string with Morse code.
}
\examples{
morse("a1!aa")
morse("a^^a")
}
