% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.LCx.R
\name{plot.LCx}
\alias{plot.LCx}
\title{Plotting method for \code{LCx} objects}
\usage{
\method{plot}{LCx}(
  x,
  xlab = "Concentration",
  ylab = "Survival probability \\n median and 95 CI",
  main = NULL,
  subtitle = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{LCx}.}

\item{xlab}{A label for the \eqn{X}-axis, by default \code{Concentration}.}

\item{ylab}{A label for the \eqn{Y}-axis, by default \code{Survival probability median and 95 CI}.}

\item{main}{A main title for the plot.}

\item{subtitle}{A subtitle for the plot}

\item{\dots}{Further arguments to be passed to generic methods.}
}
\value{
a plot of class \code{ggplot}
}
\description{
This is the generic \code{plot} S3 method for the LCx class. 
It plots the survival probability as a function of concentration.
}
\keyword{plot}
