% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survDataVarExp.R
\name{plot.survDataVarExp}
\alias{plot.survDataVarExp}
\title{Plotting method for \code{survDataVarExp} objects}
\usage{
\method{plot}{survDataVarExp}(
  x,
  xlab = "Time",
  ylab = "Number of survivors",
  main = NULL,
  one.plot = FALSE,
  facetting_level = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{survDataVarExp}}

\item{xlab}{a label for the \eqn{X}-axis, by default \code{Time}}

\item{ylab}{a label for the \eqn{Y}-axis, by default \code{Number of survivors}}

\item{main}{main title for the plot}

\item{one.plot}{if \code{TRUE}, draws all the points in one plot instead of one per \code{replicate}}

\item{facetting_level}{a vector of \code{characters} to rank \code{replicates} in the multi plot (i.e. \code{one.plot == FALSE})}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
an object of class \code{ggplot}, see function \code{\link[ggplot2]{ggplot}}
}
\description{
This is the generic \code{plot} S3 method for the \code{survDataVarC} class.
It plots the number of survivors as a function of time.
}
\keyword{plot}
