% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_post.R
\name{plot_prior_post}
\alias{plot_prior_post}
\title{Plot posteriors vs priors}
\usage{
plot_prior_post(x, ...)
}
\arguments{
\item{x}{an object used to select a method \code{plot_prior_post}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Plot posteriors vs priors of a \code{survFit} object
}
