% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoseResponse.R
\name{plotDoseResponse}
\alias{plotDoseResponse}
\title{Plot Dose-response from raw data}
\usage{
plotDoseResponse(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Plots the response as a function of the concentration (for a given
target time).
}

