% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc}
\alias{ppc}
\title{Posterior predictive check plot}
\usage{
ppc(x, ...)
}
\arguments{
\item{x}{an object used to select a method \code{ppc}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Plots posterior predictive check for \code{reproFitTT}, \code{survFitTT},
 \code{survFitTKTD}, \code{survFitCstExp} and \code{survFitVarExp} objects.
}
\details{
Depending on the class of the object \code{x} see their links. 
for class \code{reproFitTT}: \link[=ppc.reproFitTT]{ppc.reproFitTT} ;  
for class \code{survFitTT}: \link[=ppc.survFitTT]{ppc.survFitTT} ; 
for class \code{survFitTKTD}: \link[=ppc.survFitTKTD]{ppc.survFitTKTD} ;
for class \code{survFitCstExp}: \link[=ppc.survFitCstExp]{ppc.survFitCstExp} and
for class \code{survFitVarExp}: \link[=ppc.survFitVarExp]{ppc.survFitVarExp}.
}
