% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Nsurv.R
\name{predict_Nsurv}
\alias{predict_Nsurv}
\title{\code{Predict_Nsurv} method for \code{survFit} objects}
\usage{
predict_Nsurv(object, ...)
}
\arguments{
\item{object}{an object used to select a method}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
It provides the simulated number of survivors for "SD" or "IT" models under
constant or time-variable exposure.
}
