% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Nsurv_check.R
\name{predict_Nsurv_check}
\alias{predict_Nsurv_check}
\title{Checking goodness-of-fit method for \code{survFitPredict} and
 \code{survFitPredict_Nsurv} objects}
\usage{
predict_Nsurv_check(object, ...)
}
\arguments{
\item{object}{an object used to select a method \code{predict_Nsurv_check}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
It returns measures of goodness-of-fit for predictions.
}
