% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgTable.R
\name{print.msgTable}
\alias{print.msgTable}
\title{Print \code{msgTables} objects}
\usage{
\method{print}{msgTable}(x, ...)
}
\arguments{
\item{x}{an object of class \code{msgTables}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
Print in the REPL the \code{msgTables}
}
\description{
Print in the REPL the \code{msgTables}
}
