% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survDataCstExp.R
\name{summary.survDataCstExp}
\alias{summary.survDataCstExp}
\title{Summary of \code{survDataCstExp} object}
\usage{
\method{summary}{survDataCstExp}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{survDataCstExp}}

\item{quiet}{when \code{TRUE}, does not print}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
The function returns a list with the following information:
\item{NbrepTimeConc}{nb of replicates for all concentrations and time points}
\item{NbsurvTimeConc}{nb of survivors. for all concentrations and time points}
}
\description{
The generic \code{summary} S3 method for the \code{survDataCstExp} class provides
information about the structure of the data set and the experimental design.
}
\examples{
# (1) Load the data
data(cadmium1)

# (2) Create a survDataCstExp object
dat <- survData(cadmium1)

# (3) Summarize the data set
summary(dat)

}
\keyword{summary}
