% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse.R
\name{unmorse}
\alias{unmorse}
\title{Decode a string with Morse code.}
\usage{
unmorse(string, sep = "/")
}
\arguments{
\item{string}{A character vector of length 1 to decode.}

\item{sep}{Separator between code words.}
}
\value{
Decoded string.
}
\description{
Decode a string with Morse code.
}
\examples{
unmorse(".-/./.-", sep = "/")
unmorse(".-/a/.", sep = "/")
unmorse(morse("a^^a"))
}
