% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/morse.R
\docType{data}
\name{cadmium2}
\alias{cadmium2}
\title{Reproduction and survival datasets for snails exposed to cadmium during 56
days}
\format{A data frame with 612 observations of the following five variables:
\describe{ \item{\code{replicate}}{A vector of class \code{factor} with the
replicate code (\code{A} to \code{F}).} \item{\code{conc}}{A vector of
class \code{integer} with the cadmium concentrations in \eqn{\mu g.L^{-1}}.}
\item{\code{time}}{A vector of class \code{integer} with the time points
(in days from the beginning of the experiment \eqn{t = 0}).}
\item{\code{Nsurv}}{A vector of class \code{integer} with the number of
alive individuals at each time point for each concentration and each
replicate.} \item{\code{Nrepro}}{A vector of class \code{integer} with the
number of offspring at each time point for each concentration and each
replicate.} }}
\usage{
data(cadmium2)
}
\description{
Reproduction and survival datasets of chronic laboratory bioassays with
snails exposed to six concentrations of one metal contaminant (cadmium)
during 56 days. Six concentrations were tested, with six replicates per
concentration. Each replicate contained five organisms. Reproduction and
survival were monitored at 17 time points.
}
\keyword{datasets}

