% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFx.R
\name{MFx}
\alias{MFx}
\title{Predict the Multiplication Factor leading to x\% of reduction in survival
at a specific time.}
\usage{
MFx(object, ...)
}
\arguments{
\item{object}{An object used to select a method}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Generic method for \code{MFx}, a function denoted \eqn{MF(x,t)} for 
\eqn{x}\% Multiplication Factor at time \eqn{t}.
}
\details{
When class of \code{object} is \code{survFit}, see \link[=MFx.survFit]{MFx.survFit}.
}
