% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{modelData}
\alias{modelData}
\title{Create a list giving data to use in Bayesian inference.}
\usage{
modelData(x, ...)
}
\arguments{
\item{x}{An object of class \code{survData}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
A list for parameterization of priors for Bayesian inference.
}
\description{
Create a list giving data to use in Bayesian inference.
}
