% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.survDataCstExp.R
\name{modelData.survDataCstExp}
\alias{modelData.survDataCstExp}
\title{Create a data set to analyse a \code{survDataCstExp} object.}
\usage{
\method{modelData}{survDataCstExp}(x, model_type = NULL)
}
\arguments{
\item{x}{An object of class \code{survData}}

\item{model_type}{TKTD GUTS model type ('SD' or 'IT')}
}
\value{
A list for parameterization of priors for Bayesian inference.
}
\description{
Create a data set to analyse a \code{survDataCstExp} object.
}
