% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse-package.R
\docType{package}
\name{morse-package}
\alias{morse-package}
\alias{_PACKAGE}
\title{morse: Encode and decode with Morse code}
\description{
Functions to convert between Morse code and normal ASCII characters.
}
\author{
\strong{Maintainer}: Kunht Kun \email{kunhtkun@gmail.com}

}
\keyword{internal}
