\name{morse-package}
\alias{morse-package}
\alias{morse}
\docType{package}
\title{MOdelling tools for Reproduction and Survival data in Ecotoxicology}
\description{
---
}
\details{
\tabular{ll}{
Package: \tab morse\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-03\cr
License: \tab \cr
}
}
\author{
Marie Laure Delignette-Muller
<marielaure.delignettemuller@vetagro-sup.fr>, Philippe Ruiz
<philippe.ruiz@univ-lyon1.fr>, Sandrine Charles
<sandrine.charles@univ-lyon1.fr>, Wandrille Duchemin
<wandrille.duchemin@insa-lyon.fr>, Christelle Lopes
<christelle.lopes@univ-lyon1.fr>, Guillaume Kon-Kam-king
<guillaume.kon-kam-king@univ-lyon1.fr>, Philippe Veber
<philippe.veber@univ-lyon1.fr>

Maintainer: Philippe Ruiz <philippe.ruiz@univ-lyon1.fr>
}
\references{
}
\keyword{ package }
\seealso{

}
\examples{

}
