\name{morse-package}
\alias{morse-package}
\alias{morse}
\docType{package}
\title{
}
\description{
}
\details{
\tabular{ll}{
Package: \tab morse\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-03\cr
License: \tab \cr
}
}
\author{

Maintainer: 
}
\references{
}
\keyword{ package }
\seealso{

}
\examples{

}
