% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survFitPredict_Nsurv.R
\name{plot.survFitPredict_Nsurv}
\alias{plot.survFitPredict_Nsurv}
\title{Plotting method for \code{survFitPredict_Nsurv} objects}
\usage{
\method{plot}{survFitPredict_Nsurv}(x, xlab = "Time",
  ylab = "Number of survivors", main = NULL, spaghetti = FALSE,
  one.plot = FALSE, mcmc_size = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{survFitPredict_Nsurv}.}

\item{xlab}{A label for the \eqn{X}-axis, by default \code{Time}.}

\item{ylab}{A label for the \eqn{Y}-axis, by default \code{Survival rate}.}

\item{main}{A main title for the plot.}

\item{spaghetti}{If \code{TRUE}, draws a set of survival curves using
parameters drawn from the posterior distribution}

\item{one.plot}{if \code{TRUE}, draws all the estimated curves in
one plot instead of one plot per concentration.}

\item{mcmc_size}{A numerical value refering by default to the size of the mcmc in object \code{survFitPredict}.
This option is specific to \code{survFitPredict} objects for which computing time may be long.
\code{mcmc_size} can be used to reduce the number of mcmc samples in order to speed up
the computation.}

\item{\dots}{Further arguments to be passed to generic methods.}
}
\description{
This is the generic \code{plot} S3 method for the
\code{survFitPredict_Nsurv}.  It plots the fit obtained for each
concentration of chemical compound in the provided dataset.
}
\details{
The fitted curves represent the \strong{estimated survival rate} as a function
of time for each concentration.
The black dots depict the \strong{observed survival
rate} at each time point.
The function plots both 95\% binomial credible intervals for the estimated survival
rate.
If \code{spaghetti = TRUE}, the credible intervals are represented by two
dotted lines limiting the credible band, and a spaghetti plot is added to this band.
This spaghetti plot consists of the representation of simulated curves using parameter values
sampled in the posterior distribution (10\% of the MCMC chains are randomly
taken for this sample).
}
\keyword{plot}
