% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_post.R
\name{plot_prior_post}
\alias{plot_prior_post}
\title{Generic method to plot priors and posteriors.}
\usage{
plot_prior_post(x, ...)
}
\arguments{
\item{x}{an object used to select a method \code{plot_prior_post}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
an object of class \code{plot_prior_post}
}
\description{
Plot priors and posteriors of a \code{survFit} object
}
