% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_post.R
\name{plot_prior_post.survFit}
\alias{plot_prior_post.survFit}
\title{Plot posteriors vs priors}
\usage{
\method{plot_prior_post}{survFit}(x, size_sample = 1000, EFSA_name = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{survFit} used to select a method \code{plot_prior_post}}

\item{size_sample}{Size of the random generation of the distribution.
Default is \code{1e3}.}

\item{EFSA_name}{If \code{TRUE}, replace the current terminology by
the one used in the recent EFSA PPR Scientific Opinion (2018).}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
a plot of class \code{ggplot}
}
\description{
Plot posteriors vs priors of a \code{survFit} object
}
\references{
EFSA PPR Scientific Opinion (2018)
\emph{Scientific Opinion on the state of the art of Toxicokinetic/Toxicodynamic (TKTD) effect models for regulatory risk assessment of pesticides for aquatic organisms}
\url{https://www.efsa.europa.eu/en/efsajournal/pub/5377}
}
