% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoseResponse.R
\name{plotDoseResponse}
\alias{plotDoseResponse}
\title{Plot dose-response from raw data}
\usage{
plotDoseResponse(x, ...)
}
\arguments{
\item{x}{an object used to select a method \code{plotDoseRespons}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
a plot of class \code{ggplot}
}
\description{
Plots the response of the effect as a function of the concentration at a given
target time.
}
