% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.reproFitTT.R
\name{print.reproFitTT}
\alias{print.reproFitTT}
\title{Print of \code{reproFitTT} object}
\usage{
\method{print}{reproFitTT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{reproFitTT}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
print the model text and the Jags Computing information
}
\description{
This is the generic \code{print} S3 method for the \code{reproFitTT} class.
It prints the underlying JAGS model and some information on the Bayesian 
inference procedure.
}
\keyword{print}
