% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survFitTT.R
\name{print.survFitTT}
\alias{print.survFitTT}
\title{Print of \code{survFitTT} object}
\usage{
\method{print}{survFitTT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{survFitTT}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
print the model text and the Jags Computing information
}
\description{
This is the generic \code{print} S3 method for the \code{survFitTT} class.
It prints the underlying JAGS model and some information on the Bayesian 
inference procedure.
}
\keyword{print}
