% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.reproData.R
\name{summary.reproData}
\alias{summary.reproData}
\title{Summary for reproData objects}
\usage{
\method{summary}{reproData}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{reproData}}

\item{quiet}{if \code{TRUE}, does no prints}

\item{\dots}{Further arguments to be passed to generic methods.}
}
\value{
The function returns a list with the same fields than
\code{\link{summary.survData}} plus an additional one:
\item{NboffTimeConc}{nb of offspring for all concentrations and time points}
}
\description{
The generic \code{summary} S3 method for the \code{reproData} class provides
information about the structure of the dataset and the experimental design.
}
\examples{
# (1) Load the data
data(cadmium1)

# (2) Create a reproData object
cadmium1 <- reproData(cadmium1)

# (3) Summarize the dataset
summary(cadmium1)
}
\seealso{
\code{\link{reproData}}, \code{\link{summary.survData}}
}
\keyword{summary}

