% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survFitTKTD.R
\name{summary.survFitTKTD}
\alias{summary.survFitTKTD}
\title{Summary of \code{survFitTKTD} object}
\usage{
\method{summary}{survFitTKTD}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{survFitTKTD}}

\item{quiet}{when \code{TRUE}, does not print}

\item{\dots}{Further arguments to be passed to generic methods.}
}
\value{
The function returns a list with the following information:
\item{Qpriors}{quantiles of the model priors}
\item{Qposteriors}{quantiles of the model posteriors}
}
\description{
This is the generic \code{summary} S3 methode for the \code{survFitTKTD} class.
It shows the quantiles of priors and posteriors on parameters.
}
\examples{
# (1) Load the data
data(propiconazole)

# (2) Create a survData object
dat <- survData(propiconazole)

\donttest{
# (3) Run the survFitTKTD function
out <- survFitTKTD(dat)

# (4) summarize the survFitTKTD object
summary(out)
}

}
\keyword{summary}
