% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survFit.R
\name{survFit}
\alias{survFit}
\title{Method to fit a model for survival data using Bayesian inference}
\usage{
survFit(data, ...)
}
\arguments{
\item{data}{an object used to select a method 'survFit'}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Method to fit a model for survival data using Bayesian inference
}
