% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survFitTT.R
\name{survFitTT}
\alias{survFitTT}
\title{Fits a Bayesian concentration-response model for target-time survival analysis}
\usage{
survFitTT(data, ...)
}
\arguments{
\item{data}{an object used to select a method 'survFitTT'}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
an object of class \code{survFitTT}
}
\description{
Fits a Bayesian concentration-response model for target-time survival analysis
}
